packages v5

+++ src/main/java/com/example/Aula/controller/Controller.java
@@ -32,12 +32,20 @@ public class Controller {
     }
 
     @DeleteMapping({"/{id}"})
-    public void deleteClienteById(@PathVariable Long id) {
+    public String deleteClienteById(@PathVariable Long id) {
         try{
-            repository.deleteById(id);
+            Optional<Cliente> cliente = Optional.of(repository.getById(id));
+            if(cliente.isPresent()){
+                repository.deleteById(id);
+                return  "Cliente de " + id + "deletado com sucesso!";
+            }else{
+                throw new Exception("Cliente inexistente!");
+            }
+
         }catch(Exception e){
-            System.out.println("Execeção ocorreu!");
             e.printStackTrace();
+            return "O cliente de " + id + " não existe para ser deletado" +
+                    " Por favor, entre em contato com o atendimento 666 666 666 ";
         }
 
     }
